adding separata gcal_reader file
