+++ csm_web/csm_web/common.py
@@ -11,7 +11,6 @@ import pprint
 import dateutil.parser
 import pickle
 
-
 # If modifying these scopes, delete the file token.json.
 SCOPES = "https://www.googleapis.com/auth/calendar"
 TIMEZONE = "America/Los_Angeles"
@@ -111,21 +110,6 @@ def get_rooms(path):
             assert title not in rooms, f"{title} has multiple emails"
             rooms[title] = email
     return rooms
-
-
-def gcal_api_authenticate():
-    """Google Calendar Authentication. Returns the service object to be used
-    for making calls to the api."""
-    # token.json stores the user's access and refresh tokens. It is created
-    # automatically when the authorization flow completes for the first time.
-
-    store = file.Storage("token.json")
-    creds = store.get()
-    if not creds or creds.invalid:
-        flow = client.flow_from_clientsecrets("secrets/credentials.json", SCOPES)
-        creds = tools.run_flow(flow, store)
-    service = build("calendar", "v3", http=creds.authorize(Http()))
-    return service
 
 
 def book_event(event, service):
